HASlib.jl

HASlib.jl

Documentation for HASlib.jl

HASlib.calc_radius_factor — Method.

Calculate radius of a circle fitting in 4 cells cells are drawn by g = i*b1+j+b2 for i and j in -1:1

source

HASlib.get_channels — Method.

Arguments

lattice Lattice object which is used for the reciprical lattice vectors

ki wavevector of incident beam

theta angle of incident beam and the surface normal in radians

n range of generated channels

source

HASlib.hexgrid — Function.

Calculate a hexagonal grid with a radius n and a lattice constant of 1 angle between a1 and a2 = 120 deg

Arguments

n:Int64

a₁=1: base vector length 1

a₂=1: base vector length 2

pattern (n = 3):

  + + +
 + + + +
+ + + + +
 + + + +
  + + +

returns (g,gi)

g .. 2d array of lattice point coordinates

gi ... index of lattice point

source

HASlib.sort_and_cut — Method.

sort by kz² and cut away channels

Arguments

ch channels

n_open number of open channels to keep

n_closed number of closed channels to keep

source

HASlib.Generic2DLattice — Type.

Structure containing the real and the reciprical basis vectors of a surface

source

HASlib.Generic2DLattice — Method.

Arguments

a1 first lattice vector

a2 second lattice vector

Example

Generic2DLattice([1,0],[1,0])

source

HASlib.ScatteringChannels — Type.

Object containing the scattering channels and the asympotic energy

source